requirements.txt
uvicorn~=0.24.0.post1
fastapi~=0.105.0
pydantic~=2.5.2
passlib[bcrypt]~=1.7.4
motor~=3.3.2
python-jose[cryptography]~=3.3.0